Server: User: Seperated all users related endpoints to a seperate router.
