+++ vault-server/src/index.ts
@@ -4,13 +4,8 @@ import express from "express";
 import cors from "cors";
 import { json } from "body-parser";
 
-import {
-	addGameToUserVault,
-	getUser,
-	isGameExistsInUserVault,
-} from "./models/User.Model";
-
 import { router as GamesRouter } from "./Games.Router";
+import { router as UsersRouter } from "./Users.Router";
 
 const app = express();
 
@@ -18,58 +13,13 @@ app.use(cors());
 app.use(json());
 
 app.use(GamesRouter);
+app.use(UsersRouter);
 
 // TODO - replace with real DBs
 
 app.get("/check", async (_, res) => {
 	res.status(200);
 	res.json({ status: "Vault backened is OK" });
-});
-
-app.get("/users/:username", async (req, res) => {
-	const targetUsername = req.params.username;
-	console.log(`User ${targetUsername} requested from ip: ${req.ip}`);
-
-	const targetUser = getUser(targetUsername);
-
-	if (!targetUser) {
-		res.status(204);
-		res.send({
-			result: null,
-			message: `Couldn't find user ${targetUsername}`,
-		});
-		return;
-	}
-
-	console.log(`User ${targetUser?.username} found.`);
-	res.status(200);
-	res.json(targetUser);
-});
-
-app.post("/users/vault/addGame", (req, res) => {
-	const { username, gameId } = req.body;
-
-	if (!username || !gameId) {
-		res.status(400);
-		res.send({
-			error: `invalid username ${username} or game id ${gameId}`,
-		});
-		return;
-	}
-
-	if (isGameExistsInUserVault(username, gameId)) {
-		res.status(400);
-		res.send({
-			code: `GAME_ALREADY_EXISTS`,
-			error: `game #${gameId} already exists in user ${username}'s vault.`,
-		});
-		return;
-	}
-
-	const updatedVault = addGameToUserVault(username, gameId);
-
-	res.status(200);
-	res.json({ message: "OK", updatedVault: updatedVault });
 });
 
 async function init() {
